+++ examples/flutter_hydrated_login/lib/bloc/auth/auth_state.dart
@@ -1,5 +1,4 @@
 import 'package:equatable/equatable.dart';
-import 'package:flutter_hydrated_login/bloc/auth/bloc.dart';
 import 'package:flutter_hydrated_login/models/user.dart';
 
 abstract class AuthState extends Equatable {
@@ -7,11 +6,9 @@ abstract class AuthState extends Equatable {
   List<Object> get props => [];
 }
 
-
-
 class AuthEmpty extends AuthState {}
 
-class AuthUser extends AuthState{
+class AuthUser extends AuthState {
   final User user;
 
   AuthUser(this.user);
