+++ examples/flutter_hydrated_login/lib/bloc/auth/bloc.dart
@@ -1,5 +1,3 @@
-
-
 export 'auth_bloc.dart';
 export 'auth_event.dart';
 export 'auth_state.dart';
