+++ examples/flutter_hydrated_login/lib/bloc/bloc_watcher.dart
