+++ examples/flutter_hydrated_login/lib/bloc/login/login_event.dart
@@ -1,7 +1,4 @@
-
-
 import 'package:equatable/equatable.dart';
-import 'login_state.dart';
 
 enum AuthStatus { success, failed, login }
 
@@ -9,7 +6,6 @@ abstract class LoginEvent extends Equatable {
   @override
   List<Object> get props => [];
 }
-
 
 class EmailChange extends LoginEvent {
   final String email;
