+++ examples/flutter_hydrated_login/lib/models/user.dart
@@ -1,5 +1,3 @@
-import 'dart:convert';
-
 import 'package:equatable/equatable.dart';
 
 class User extends Equatable {
