+++ examples/flutter_hydrated_login/lib/routes/app_pages.dart
@@ -6,8 +6,8 @@ import 'package:flutter_hydrated_login/screens/login/login_page.dart';
 class AppPages {
   static Map<String, Widget Function(BuildContext)> routes() {
     return {
-      AppRoutes.HOME : (context) => HomePage(),
-      AppRoutes.LOGIN : (context) => LoginPage()
+      AppRoutes.HOME: (context) => HomePage(),
+      AppRoutes.LOGIN: (context) => LoginPage()
     };
   }
 }
