+++ examples/flutter_hydrated_login/lib/routes/app_routes.dart
@@ -1,7 +1,4 @@
-
-
 class AppRoutes {
-
   static const String LOGIN = '/login';
   static const String HOME = '/home';
 }
