+++ examples/flutter_hydrated_login/lib/screens/login/password_form.dart
@@ -2,20 +2,24 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_hydrated_login/bloc/login/bloc.dart';
 
+
 class PasswordFormTextField extends StatelessWidget {
+  
+  
   @override
   Widget build(BuildContext context) {
+    
     return Padding(
         padding: const EdgeInsets.symmetric(horizontal: 16.0),
-        child:BlocBuilder<LoginBloc, LoginState>(buildWhen: (previous, current) {
+        child:
+            BlocBuilder<LoginBloc, LoginState>(buildWhen: (previous, current) {
           if (previous is MyFormChange && current is MyFormChange) {
             return previous.password != current.password;
           }
+          return false;
         }, builder: (_, state) {
-          var item;
           var errorTextState;
           if (state is MyFormChange) {
-            item = state.password;
             errorTextState = state.passwordValidation();
           }
           return TextField(
